Replace jest with node built-in test runner
