--- jest.post.config.ts
@@ -1,20 +1,0 @@
-import type { Config } from '@jest/types';
-
-// Sync object
-const config: Config.InitialOptions = {
-  verbose: true,
-  clearMocks: true,
-  moduleFileExtensions: ['js', 'ts'],
-  transform: {
-    '^.+\\.(t|j)sx?$': [
-      'esbuild-jest-transform',
-      {
-        'target': 'node20',
-        'packages': 'external',
-      },
-    ],
-  },
-  extensionsToTreatAsEsm: ['.ts'],
-  testMatch: ['**/__tests__/postbuild/**/*.test.ts'],
-};
-module.exports = config;
